refactoring: host app의 App 관련 파일 디렉토리 변경
./src/... -> ./src/App/...
