+++ frontend/host-app/src/App/App.js
